.history/backend/requirements_20250319184504.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise
gunicorn
django-cors-headers
python-dotenv
psycopg2-binary
dj-database-url